requirements.txt
torch
matplotlib
numpy
pillow
scikit-image
tqdm
gradio
